Fix and refactor LeadForm
Renamed delete_cached_file to delete_file

We're flagging if a file has error for the purposes of determining if
we need to cache it with the has_error flag, which is global. If there
were errors from other fields before the image is checked, the image
will not get cached. Instead, we use a local img_error flag.